cos/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
slack-sdk>=3.33.0
twilio>=9.3.0
